+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+16.9.0 - 2020-08-05
+=====================================================
+
+* Require expected Delayed Job worker count be derived from an ENV variable to
+  ensure that monit can be synced with Marty to manage Delayed Job workers.
+* Various small fixes related to interacting with Delayed Job workers through
+  the Marty UI
+
 16.8.0 - 2020-07-16
 =====================================================